# docs/core/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SSTVMode": {
        "description": "Supported SSTV modes.",
        "enum": [
          "MartinM1",
          "MartinM2",
          "ScottieS1",
          "ScottieS2",
          "ScottieDX",
          "Robot36",
          "Robot72",
          "PD90",
          "PD120",
          "PD180",
          "PD240",
          "WraaseSC2_180"
        ],
        "title": "SSTVMode",
        "type": "string"
      },
...
      "TransmitRequest": {
        "description": "Request to transmit an SSTV image.",
        "properties": {
          "callsign": {
            "anyOf": [
              {
                "maxLength": 20,
                "minLength": 3,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Operator callsign (overlaid on image)",
            "title": "Callsign"
          },
          "device_id": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Audio output device ID (null for system default)",
            "title": "Device Id"
          },
          "image_path": {
            "description": "Absolute path to image file",
            "maxLength": 4096,
            "minLength": 1,
            "title": "Image Path",
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/SSTVMode",
            "description": "SSTV mode for transmission"
          },
          "serial_port": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Serial port for PTT control (e.g., /dev/ttyUSB0, COM3)",
            "title": "Serial Port"
          },
          "vox_enabled": {
            "default": false,
            "description": "Use VOX (silence preamble) instead of PTT",
            "title": "Vox Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "image_path",
          "mode"
        ],
        "title": "TransmitRequest",
        "type": "object"
      },
      "TransmitResponse": {
        "description": "Response from starting a transmission.",
        "properties": {
          "estimated_duration_seconds": {
            "description": "Estimated transmission duration",
            "minimum": 0.0,
            "title": "Estimated Duration Seconds",
            "type": "number"
          },
          "started_at": {
            "description": "Transmission start timestamp (UTC)",
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/TransmitState",
            "description": "Current transmission state"
          },
          "tx_id": {
            "description": "Unique transmission identifier",
            "format": "uuid",
            "title": "Tx Id",
            "type": "string"
          },
          "websocket_url": {
            "description": "WebSocket URL for real-time updates",
            "title": "Websocket Url",
            "type": "string"
          }
        },
        "required": [
          "tx_id",
          "state",
          "websocket_url",
          "estimated_duration_seconds",
          "started_at"
        ],
        "title": "TransmitResponse",
        "type": "object"
      },
      "TransmitSmartReplyRequest": {
        "description": "Request to transmit Smart Reply.",
        "properties": {
          "device_id": {
            "description": "Audio output device ID",
            "title": "Device Id",
            "type": "string"
          },
          "mode": {
            "description": "SSTV mode (ScottieS1, MartinM1, Robot36)",
            "title": "Mode",
            "type": "string"
          },
          "ptt_method": {
            "description": "PTT method (none, serial, vox)",
            "title": "Ptt Method",
            "type": "string"
          }
        },
        "required": [
          "mode",
          "device_id",
          "ptt_method"
        ],
        "title": "TransmitSmartReplyRequest",
        "type": "object"
      },
      "TransmitSmartReplyResponse": {
        "description": "Response from Smart Reply transmission.",
        "properties": {
          "status": {
            "description": "Transmission status",
            "title": "Status",
            "type": "string"
          },
          "tx_id": {
            "description": "Transmit session ID",
            "format": "uuid",
            "title": "Tx Id",
            "type": "string"
          }
        },
        "required": [
          "tx_id",
          "status"
        ],
        "title": "TransmitSmartReplyResponse",
        "type": "object"
      },
      "TransmitState": {
        "description": "Transmit session states.",
        "enum": [
          "pending",
          "ptt_engaged",
          "transmitting",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "TransmitState",
        "type": "string"
      },
      "TransmitStatusResponse": {
        "description": "Transmission status.",
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "elapsed_seconds": {
            "default": 0.0,
            "description": "Elapsed transmission time",
            "minimum": 0.0,
            "title": "Elapsed Seconds",
            "type": "number"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Error message if state is 'failed'",
            "title": "Error"
          },
          "estimated_duration_seconds": {
            "description": "Total estimated duration",
            "minimum": 0.0,
            "title": "Estimated Duration Seconds",
            "type": "number"
          },
          "image_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Image ID if saved to library",
            "title": "Image Id"
          },
          "mode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SSTVMode"
              },
              {
                "type": "null"
              }
            ]
          },
          "progress_percent": {
            "default": 0.0,
            "description": "Transmission progress (0-100)",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress Percent",
            "type": "number"
          },
          "scanlines_transmitted": {
            "default": 0,
            "description": "Number of scanlines transmitted",
            "minimum": 0.0,
            "title": "Scanlines Transmitted",
            "type": "integer"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/TransmitState"
          },
          "tx_id": {
            "format": "uuid",
            "title": "Tx Id",
            "type": "string"
          }
        },
        "required": [
          "tx_id",
          "state",
          "estimated_duration_seconds",
          "started_at"
        ],
        "title": "TransmitStatusResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "SSTV (Slow-Scan Television) encode/decode engine",
    "title": "SSTeVe API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/smart_reply/transmit/{preview_id}": {
      "post": {
        "description": "Transmit Smart Reply preview image.\n\nArgs:\n    preview_id: Preview ID from generate endpoint\n    request: Transmission parameters (mode, device, PTT)\n    db: Database session\n\nReturns:\n    Transmission session information\n\nRaises:\n    HTTPException: If preview not found or transmission fails",
        "operationId": "transmit_smart_reply_api_v1_smart_reply_transmit__preview_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "preview_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Preview Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransmitSmartReplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransmitSmartReplyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transmit Smart Reply",
        "tags": [
          "Smart Reply",
          "smart_reply"
        ]
      }
    },
    "/api/v1/transmit": {
      "post": {
        "description": "Start a new SSTV transmission.\n\nEncodes the image, engages PTT (if configured), and transmits the SSTV signal.\nThe transmission duration depends on the SSTV mode selected.\n\nReturns:\n    TransmitResponse with transmission ID and WebSocket URL for real-time updates.\n\nRaises:\n    409 Conflict: If another decode/transmit session is already active (half-duplex)\n    404 Not Found: If image file doesn't exist\n    400 Bad Request: If image format is unsupported or invalid",
        "operationId": "start_transmit_api_v1_transmit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransmitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Transmit",
        "tags": [
          "Transmit",
          "transmit"
        ]
      }
    },
    "/api/v1/transmit/cancel/{tx_id}": {
      "post": {
        "description": "Cancel an active transmission.\n\nReleases PTT and stops transmitting. The partial transmission will be aborted.\n\nRaises:\n    404 Not Found: If transmission doesn't exist\n    409 Conflict: If transmission is already in a terminal state",
        "operationId": "cancel_transmit_api_v1_transmit_cancel__tx_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "tx_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tx Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Transmit",
        "tags": [
          "Transmit",
          "transmit"
        ]
      }
    },
    "/api/v1/transmit/status/{tx_id}": {
      "get": {
        "description": "Get the current status of a transmission.\n\nReturns detailed progress information including:\n- Current state (pending, ptt_engaged, transmitting, completed, etc.)\n- Progress percentage and scanlines transmitted\n- Elapsed time and estimated total duration\n\nRaises:\n    404 Not Found: If transmission doesn't exist",
        "operationId": "get_transmit_status_api_v1_transmit_status__tx_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "tx_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tx Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransmitStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Transmit Status",
        "tags": [
          "Transmit",
          "transmit"
        ]
      }
    }
  }
}